Rector code for easier maintainability
