+++ app/src/main/java/com/github/se/polyfit/ui/components/PictureDialog.kt
